squash! Saving intermediate work
